Initial and VERY experimental database result iterators, so far it only works on  - MySQL  - MySQLi
I haven't looked at SQLite yet, but PDO needs an emulation layer which is most likely 
to be very costly, but lets see if its possible to cook something reliable up.

This is a step towards the pagination helper class, as the idea is to support every 
class that implements the \Iterator interface, although it might be impossible its a 
decent feature nonetheless to the database layer.