Minor cosmetic improvements to \Tuxxedo\Debug\Backtrace Removed the UTF-8 convertion when editing a template, which in some odd cases caused blank 'sources' Removed old global constants:  - TUXXEDO_DEBUG  - TUXXEDO_SELF  - TUXXEDO_USERAGENT Removed the $debug_trim parameter on tuxxedo_trim_path() Removed calls to ini_set(), except the timeout in the MySQLi driver Removed the $registry->configuration reference, use Registry::getConfiguration() Removed the page_print() function
It is the plan to clean up some more of these global space functions and constants, while these were the 'easiest' and most unused ones, they were easy to get rid of.

The remaining ones can be split into 3 categories:
* Handler functions (Error, Exception, Shutdown)
* Error handling system (doc_error, cli_error, etc)
* Misc. (date, trim, redirect, etc)

How we are gonna cope with the TUXXEDO_PREFIX I still don't know, or if we even are gonna change that, but it would be nice to bring consistency to the API before the bundle merge in 1.3.